Add files via upload
updated test #1 - includes statistical calculations for all the different layers of all the models. stores the output for the ideal and analog model of each layer for each architecture, calculates mean and std for each layer of each model and plots the mean with the std as an error bar.